requirements.txt
aiogram==2.14.3
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.3
idna==3.2
multidict==5.1.0
numpy==1.21.1
pydub==0.25.1
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
ShazamAPI==0.0.2
typing-extensions==3.10.0.0
urllib3==1.26.6
yarl==1.6.3
youtube-dl==2021.6.6
